+++ provider/resources.go
@@ -19,14 +19,12 @@ import (
 	"path/filepath"
 	"unicode"
 
+	"github.com/dnsimple/terraform-provider-dnsimple/dnsimple"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 	"github.com/pulumi/pulumi-dnsimple/provider/v3/pkg/version"
 	"github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tfbridge"
-	shim "github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tfshim"
 	shimv1 "github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tfshim/sdk-v1"
-	"github.com/pulumi/pulumi/sdk/v3/go/common/resource"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/tokens"
-	"github.com/terraform-providers/terraform-provider-dnsimple/dnsimple"
 )
 
 // all of the token components used below.
@@ -55,28 +53,20 @@ func makeResource(mod string, res string) tokens.Type {
 	return makeType(mod+"/"+fn, res)
 }
 
-// preConfigureCallback is called before the providerConfigure function of the underlying provider.
-// It should validate that the provider can be configured, and provide actionable errors in the case
-// it cannot be. Configuration variables can be read from `vars` using the `stringValue` function -
-// for example `stringValue(vars, "accessKey")`.
-func preConfigureCallback(vars resource.PropertyMap, c shim.ResourceConfig) error {
-	return nil
-}
-
 // Provider returns additional overlaid schema and metadata associated with the provider..
 func Provider() tfbridge.ProviderInfo {
 	p := shimv1.NewProvider(dnsimple.Provider().(*schema.Provider))
 
 	prov := tfbridge.ProviderInfo{
-		P:                    p,
-		Name:                 "dnsimple",
-		Description:          "A Pulumi package for creating and managing dnsimple cloud resources.",
-		Keywords:             []string{"pulumi", "dnsimple"},
-		License:              "Apache-2.0",
-		Homepage:             "https://pulumi.io",
-		Repository:           "https://github.com/pulumi/pulumi-dnsimple",
-		PreConfigureCallback: preConfigureCallback,
-		Config:               map[string]*tfbridge.SchemaInfo{},
+		P:           p,
+		Name:        "dnsimple",
+		Description: "A Pulumi package for creating and managing dnsimple cloud resources.",
+		Keywords:    []string{"pulumi", "dnsimple"},
+		License:     "Apache-2.0",
+		Homepage:    "https://pulumi.io",
+		Repository:  "https://github.com/pulumi/pulumi-dnsimple",
+		Config:      map[string]*tfbridge.SchemaInfo{},
+		GitHubOrg:   "dnsimple",
 		Resources: map[string]*tfbridge.ResourceInfo{
 			"dnsimple_record": {Tok: makeResource(mainMod, "Record"),
 				Fields: map[string]*tfbridge.SchemaInfo{
@@ -94,11 +84,12 @@ func Provider() tfbridge.ProviderInfo {
 					},
 				},
 			},
+			"dnsimple_email_forward": {Tok: makeResource(mainMod, "EmailForward")},
 		},
 		DataSources: map[string]*tfbridge.DataSourceInfo{},
 		JavaScript: &tfbridge.JavaScriptInfo{
 			Dependencies: map[string]string{
-				"@pulumi/pulumi": "^3.0.0-alpha.0",
+				"@pulumi/pulumi": "^3.0.0",
 			},
 			DevDependencies: map[string]string{
 				"@types/node": "^10.0.0", // so we can access strongly typed node definitions.
@@ -112,7 +103,7 @@ func Provider() tfbridge.ProviderInfo {
 		},
 		Python: &tfbridge.PythonInfo{
 			Requires: map[string]string{
-				"pulumi": ">=3.0.0a1,<4.0.0",
+				"pulumi": ">=3.0.0,<4.0.0",
 			},
 		},
 		Golang: &tfbridge.GolangInfo{
@@ -126,7 +117,7 @@ func Provider() tfbridge.ProviderInfo {
 		},
 		CSharp: &tfbridge.CSharpInfo{
 			PackageReferences: map[string]string{
-				"Pulumi":                       "3.*-*",
+				"Pulumi":                       "3.*",
 				"System.Collections.Immutable": "1.6.0",
 			},
 			Namespaces: map[string]string{