+++ sdk/dotnet/Record.cs
@@ -60,13 +60,19 @@ namespace Pulumi.DNSimple
     /// 
     /// ## Import
     /// 
-    /// DNSimple resources can be imported using their domain name and numeric ID, e.g.
+    /// DNSimple resources can be imported using their parent zone name (domain name) and numeric record ID. **Importing record example.com with record ID 1234**
     /// 
     /// ```sh
     ///  $ pulumi import dnsimple:index/record:Record resource_name example.com_1234
     /// ```
     /// 
-    ///  The numeric ID can be found in the URL when editing a record on the dnsimple web dashboard.
+    ///  **Importing record www.example.com with record ID 1234**
+    /// 
+    /// ```sh
+    ///  $ pulumi import dnsimple:index/record:Record resource_name example.com_1234
+    /// ```
+    /// 
+    ///  The record ID can be found in the URL when editing a record on the DNSimple web dashboard.
     /// </summary>
     [DNSimpleResourceType("dnsimple:index/record:Record")]
     public partial class Record : Pulumi.CustomResource