+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 ## HEAD (Unreleased)
-* Depend on Pulumi 3.0, which includes improvements to Python resource arguments and key translation, Go SDK performance,
-  Node SDK performance, general availability of Automation API, and more.
+* Upgrade to v0.5.1 of the DNSimple Terraform Provider
 
 ---
+
+## 3.0.0 (2021-04-19)
+* Depend on Pulumi 3.0, which includes improvements to Python resource arguments and key translation, Go SDK performance,
+  Node SDK performance, general availability of Automation API, and more.
 
 ## 2.7.0 (2021-04-12)
 * Upgrade to pulumi-terraform-bridge v2.23.0